Move high-level wrappers to new module
